+++ packages/@tailwindcss-upgrade/src/codemods/template/migrate-arbitrary-utilities.ts
@@ -1,12 +1,15 @@
 import { printModifier, type Candidate } from '../../../../tailwindcss/src/candidate'
 import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
 import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
+import {
+  computeUtilitySignature,
+  preComputedUtilities,
+} from '../../../../tailwindcss/src/signatures'
 import type { Writable } from '../../../../tailwindcss/src/types'
 import { DefaultMap } from '../../../../tailwindcss/src/utils/default-map'
 import { dimensions } from '../../../../tailwindcss/src/utils/dimensions'
 import * as ValueParser from '../../../../tailwindcss/src/value-parser'
 import { baseCandidate, parseCandidate } from './candidates'
-import { computeUtilitySignature, preComputedUtilities } from './signatures'
 
 const baseReplacementsCache = new DefaultMap<DesignSystem, Map<string, Candidate>>(
   () => new Map<string, Candidate>(),
